♻ refactor(api): simplifica a inicialização da API
