+++ api/index.ts
@@ -1,15 +1,11 @@
-import express from 'express';
+import app from './src/app';
+import connectToDatabase from './src/db/connection';
 
-// import server from './server';
-
-// server.startServer();
-
-const server = express();
-
-server.get('/', (_req, res) => {
-  res.sendStatus(200);
-});
-
-server.listen(3001, () => {
-  console.log('API rodando na porta 3001!');
+connectToDatabase().then(() => {
+  app.listen(
+    3001,
+    () => console.log(
+      'Api running on port 3001.',
+    ),
+  );
 });
