+++ api/src/tests/mocks/TodoMocks.ts
@@ -1,0 +1,10 @@
+import { TodoSchema } from '../../app/interfaces/TodoSchema';
+
+export const validTodo: TodoSchema = {
+  id: 'someNiceID',
+  description: 'Fazer testes unitários',
+  status: 'A fazer',
+  createdAt: new Date(),
+};
+
+export const invalidTodo = { invalid: true } as any;
